Add lambda handler click
