+++ app/src/main/res/layout/view_holder_movie.xml
@@ -11,8 +11,9 @@
         android:id="@+id/iv_movie"
         android:layout_width="wrap_content"
         android:layout_height="300dp"
-        android:layout_marginStart="4dp"
-        android:layout_marginBottom="8dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="4dp"
+        android:layout_marginBottom="4dp"
         android:background="@drawable/roundrect"
         android:contentDescription="@string/todo10"
         android:scaleType="fitStart"
